+++ cynic-codegen/src/generic_param.rs
@@ -29,10 +29,12 @@ impl GenericParameter {
 /// Our generic parameters need constraints - this enum specifies what they
 /// should be.
 pub enum GenericConstraint {
-    /// An enum type constraint: `where T: Enum<SomeEnumMarkerStruct>
+    /// An enum type constraint: `where T: Enum<SomeTypeLockMarkerStruct>
     Enum(Ident),
     /// An input object constraint: `where T: InputObject<SomeInputObjectMarkerStruct>
     InputObject(Ident),
+    /// A scalar object constraint: `where T: Scalar<X>
+    Scalar(TypePath),
 }
 
 impl GenericConstraint {
@@ -49,6 +51,11 @@ impl GenericConstraint {
                 let type_path = TypePath::concat(&[path_to_markers, ident.clone().into()]);
 
                 quote! { ::cynic::InputObject<#type_path> + ::cynic::SerializableArgument }
+            }
+            GenericConstraint::Scalar(scalar_path) => {
+                let type_path = TypePath::concat(&[path_to_markers, scalar_path.clone()]);
+
+                quote! { ::cynic::Scalar<#type_path> + ::cynic::SerializableArgument }
             }
         }
     }