+++ cynic/src/bin/simple.rs
@@ -5,8 +5,6 @@ fn main() {
 }
 
 mod query_dsl {
-    type Json = serde_json::Value;
-
     cynic::query_dsl!("src/bin/simple.graphql");
 }
 
@@ -137,11 +135,14 @@ mod test {
     // A custom scalars.
     pub struct DateTime {}
 
-    impl cynic::Scalar for DateTime {
-        fn decode(_: &serde_json::Value) -> Result<Self, json_decode::DecodeError> {
-            Ok(DateTime {})
+    impl cynic::Scalar<super::query_dsl::Json> for DateTime {
+        type Codable = String;
+
+        fn from_codable(_inner: Self::Codable) -> Result<Self, ::cynic::DecodeError> {
+            todo!()
         }
-        fn encode(&self) -> Result<serde_json::Value, ::cynic::SerializeError> {
+
+        fn to_codable(&self) -> Result<&Self::Codable, ::cynic::SerializeError> {
             todo!()
         }
     }