+++ cynic/src/id.rs
@@ -22,12 +22,14 @@ impl<T: Into<String>> From<T> for Id {
     }
 }
 
-impl crate::Scalar for Id {
-    fn decode(value: &serde_json::Value) -> Result<Self, json_decode::DecodeError> {
-        String::decode(value).map(Into::into)
+impl crate::Scalar<Id> for Id {
+    type Codable = String;
+
+    fn from_codable(s: String) -> Result<Self, json_decode::DecodeError> {
+        Ok(s.into())
     }
 
-    fn encode(&self) -> Result<serde_json::Value, SerializeError> {
-        self.0.encode()
+    fn to_codable(&self) -> Result<&Self::Codable, SerializeError> {
+        Ok(&self.0)
     }
 }